requirements.txt
astroid==2.4.2
autopep8==1.5.3
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
gcloud==0.18.3
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.0
oauth2client==4.1.3
pandas==1.0.5
pgeocode==0.2.1
protobuf==3.12.2
py==1.8.2
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retry==0.9.2
rope==0.17.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
